requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Deep Learning & RL
torch>=2.0.0      # We will use PyTorch
gymnasium>=0.28.1 # The modern replacement for OpenAI Gym

# Experiment Tracking
wandb>=0.15.0     # Weights & Biases
pyyaml>=6.0       # For reading config files
tqdm>=4.65.0      # Progress bars

# Development & Testing
pytest>=7.3.0
black             # Code formatter
flake8            # Linter
jupyter           # For notebooks